Added endpoint for flight end

+++ Request Dispatcher/Messages/FlightMessage.cs
@@ -8,9 +8,9 @@ namespace Request_Dispatcher.Messages
         public int TeamID { get; set; }
         public string FlightID { get; set; }
         public int PlatoonID { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
         public DateTime BeginTime { get; set; }
         public string? Comment { get; set; }
